+++ test/unit/SchemaPack.test.js
@@ -8,7 +8,6 @@ const expect = require('chai').expect,
   REMOTE_REFS = 'test/data/separatedFiles/remoteRefs',
   COUNTING_SEPARATED_FOLDER = '../data/separatedFiles/counting',
   WIKI_20_FOLDER = '../data/separatedFiles/wiki20',
-  MULTIPLE_ROOT = '../data/separatedFiles/multipleRoot',
   fs = require('fs'),
   getAllTransactionsFromCollection = require('../../lib/utils/getAllTransactions').getAllTransactionsFromCollection,
   async = require('async'),
@@ -1167,32 +1166,6 @@ describe('SchemaPack detectRelatedFiles', async function () {
 
 describe('SchemaPack detectRootFiles', function () {
 
-  it('should return error when input is an empty object', async function () {
-    try {
-      const schemaPack = new SchemaPack({}, {});
-      await schemaPack.detectRootFiles();
-    }
-    catch (error) {
-      expect(error).to.not.be.undefined;
-      expect(error.message).to.equal('Input object must have "type" and "data" information');
-    }
-  });
-
-  it('should return error when input data is an empty array', async function () {
-    try {
-      const
-        schemaPack = new SchemaPack({
-          type: 'multiFile',
-          data: ''
-        }, {});
-      await schemaPack.detectRootFiles();
-    }
-    catch (error) {
-      expect(error).to.not.be.undefined;
-      expect(error.message).to.equal('"Data" parameter should be provided');
-    }
-  });
-
   it('should return one root 1.1 correctly without specificationVersion provided', async function () {
     const service = path.join(
         __dirname,
@@ -1408,131 +1381,6 @@ describe('SchemaPack detectRootFiles', function () {
     expect(result.output.data.length).to.equal(0);
     expect(result.output.specification.version).to.equal('1.1');
     expect(result.output.type).to.be.equal('rootFiles');
-  });
-
-  it('should return 2 root 1.1 correctly', async function () {
-    const service = path.join(
-        __dirname,
-        MULTIPLE_ROOT,
-        '/CountingCategoryService.wsdl'
-      ),
-      service2 = path.join(
-        __dirname,
-        MULTIPLE_ROOT,
-        '/CountingCategoryServiceCopy.wsdl'
-      ),
-      types = path.join(
-        __dirname,
-        COUNTING_SEPARATED_FOLDER,
-        '/CountingCategoryData.xsd'
-      );
-    let serviceContent = fs.readFileSync(service, 'utf8'),
-      serviceContent2 = fs.readFileSync(service2, 'utf8'),
-      typesContent = fs.readFileSync(types, 'utf8'),
-      input = {
-        type: 'multiFile',
-        data: [
-          {
-            path: '/CountingCategoryService.wsdl',
-            content: serviceContent
-          },
-          {
-            path: '/CountingCategoryData.xsd',
-            content: typesContent
-          },
-          {
-            path: '/CountingCategoryServiceCopy.wsdl',
-            content: serviceContent2
-          }
-        ]
-      },
-      schemaPack = new SchemaPack(input, {}),
-      result = await schemaPack.detectRootFiles();
-    expect(result.result).to.be.true;
-    expect(result.output.data.length).to.equal(2);
-    expect(result.output.specification.version).to.equal('1.1');
-    expect(result.output.type).to.be.equal('rootFiles');
-  });
-
-  it('should propagate one error correctly', async function () {
-    const service = path.join(
-      __dirname,
-      COUNTING_SEPARATED_FOLDER,
-      '/CountingCategoryService.wsdl'
-    );
-    let serviceContent = fs.readFileSync(service, 'utf8'),
-      input = {
-        type: 'multiFile',
-        specificationVersion: '1.1',
-        data: [
-          {
-            path: '',
-            content: serviceContent
-          }
-        ]
-      };
-    try {
-      const schemaPack = new SchemaPack(input, {});
-      await schemaPack.detectRootFiles();
-    }
-    catch (ex) {
-      expect(ex.message).to.equal('"Path" of the data element should be provided');
-    }
-  });
-
-  it('should not read content from FS when is not present', async function () {
-    let input = {
-        type: 'multiFile',
-        data: [
-          {
-            path: '/CountingCategoryService.wsdl'
-          },
-          {
-            path: '/CountingCategoryData.xsd'
-          }
-        ]
-      },
-      schemaPack = new SchemaPack(input, {}),
-      result = await schemaPack.detectRootFiles();
-    expect(result.result).to.be.true;
-    expect(result.output.data.length).to.equal(0);
-    expect(result.output.specification.version).to.equal('1.1');
-    expect(result.output.type).to.be.equal('rootFiles');
-  });
-
-  it('should return error when "type" parameter is not sent', async function () {
-    const service = path.join(
-        __dirname,
-        COUNTING_SEPARATED_FOLDER,
-        '/CountingCategoryService.wsdl'
-      ),
-      types = path.join(
-        __dirname,
-        COUNTING_SEPARATED_FOLDER,
-        '/CountingCategoryData.xsd'
-      );
-    let serviceContent = fs.readFileSync(service, 'utf8'),
-      typesContent = fs.readFileSync(types, 'utf8'),
-      input = {
-        data: [
-          {
-            path: '/CountingCategoryService.wsdl',
-            content: serviceContent
-          },
-          {
-            path: '/CountingCategoryData.xsd',
-            content: typesContent
-          }
-        ]
-      };
-    try {
-      const schemaPack = new SchemaPack(input, {});
-      await schemaPack.detectRootFiles();
-    }
-    catch (error) {
-      expect(error).to.not.be.undefined;
-      expect(error.message).to.equal('"Type" parameter should be provided');
-    }
   });
 
 });
